requirements.txt
fastapi
uvicorn
pyyaml
google-auth
google-auth-oauthlib
google-api-python-client
pydantic
